Close SF patch #481914:
The included patch enables the Delphi colorizer
for .pas and .dpr files. Most code should be delphi or
some form of Turbo Pascal equivilent syntax if it uses
the .pas. Currently ViewCVS has this mapped to the
ANSI standard 'pascal' colorizer. I checked with
Michael van Canneyt, the maintainer of both the Pascal
and Delphi colorizers in Enscript and he recommended
this change.


git-svn-id: http://viewvc.tigris.org/svn/viewvc/trunk@318 8cb11bc2-c004-0410-86c3-e597b4017df7