direasy.ezt is a simple variation based on directory.ezt. Recently Stephen Lamm wrote: > Have file links go to the latest revision instead of the cvs log. > When I click on the file name, I expect to get that file, and not > some meta information about it. Another column next the name could > have a link like "(Log)" to get that information. That would be > more explicit. I find it strange to click on the revision number > to get a file because I typically do not think of files in terms > of revision numbers. (The revision number probably should not have > a link on it at all.)
In this template the revision number carries a link to the Log
page  (so the meaning of both columns are simply swapped).
Additionally (if you enable cvsgraph) the text string "(graph)" is replaced
by a small 16x16 pixel icon.

I experimented with the suggestion to add the word '(Log)' to each
filename but in the end I think that both (graph) and (log)
waste a lot of valueable horizontal screen space.

If you want to try this template out, you can simply start
	./standalone.py --gui
in the viewcvs directory and select the radio button labeled
with 'direasy.ezt'.


git-svn-id: http://viewvc.tigris.org/svn/viewvc/trunk@322 8cb11bc2-c004-0410-86c3-e597b4017df7