Modified ccvs so that it conforms to the new API by Greg. Modified a bit the API also:  * Added a Versitem class from which both Versfile and Versitem inherite.  * Added a show_CVSROOT constructor parameter in the CVSRepository.
Now that I am looking over to this code after some time, There are number
of concerns/interrogation that are coming to my mind:
 - Why not go into some more abstraction/polymorphism with the items  ( e.g.
Versfiles and Versdirs):
	shouldn't Versitems  have only one common getitemlist methods ?
	that would return Versitems
( and not two "getsubdirs" and "getfiles" methods...)

 - Quite obviously there are cyclic references in this library.
     This will flaw garbage collection.
    e.g. Versfile points to Repository. and so on...
  According to me deleting a Repository should in turn, result in the destruction of
 its subitems.
 - Another point of concern, we will have to deal with is
multithreading...  One of the current problem of viewcvs is that it does not
have multithreading. ( being a cgi it's kind of normal)  Now, having
multithreading capabilities in vclib would definitely leverage the
difficulty of porting  viewcvs to a servlet-like behaviour. A servlet-like
behaviour would be a big win for speed, according to me.

 - A very important ( if not vital) feature is lacking:
    You cannot select a tag to view in the repository ...
 - There is a problem with subversion ( named svn, from now on...):
svn, as you probably know, represent the evolution of a file hierarchy over time and not the
 evolution of files in hierarchy ( as in CVS). This API does not work with this kind of behaviour.


git-svn-id: http://viewvc.tigris.org/svn/viewvc/trunk@526 8cb11bc2-c004-0410-86c3-e597b4017df7